Added updated version of sobreposicoes python script
This one only requires "requests" to be installed and does everything
via SIGARRA's API (the previous version exclusively used web scraping
via selenium).

It also collects the password "safely" (instead of via a command line
argument).

A simple test that was ran using the same username and password for the
IOPE, LGP and CPAR course units for the school year of 2019 (-2020).
The results were the same but this version took 3-4 seconds (depends on
how fast you are at pasting the password in) versus the old version's 1
minute and 38 seconds.
Hooray for speed!

The only thing the new version does not do that the old one does it
checking all of the subsets.
For example, for the UCs given above the new version only shows the
intersection of the 3. However, the old version shows the intersection
of each pair and also of the 3.

This seems like a kind of useless feature but might be interesting to
have it as an argument switch or something just in case it's useful.
However, due to how much faster the new version is, you can just run it
repeatedly to get the same results :grin: